Move threadDelay for gcThread at the beginning
This commit moves the `threadDelay` present in the `gcThread` function
at the beginning of the `go` aux rather than at the end. This has the
effect that the first GC happens _after_ the delay time and not
immediately, which might make some tests unpredictable if we get unlucky
and some job ids are removed from the underlying map before we can
inspect their state.