[FIX] Graph of Order 2 A or 2 B ? Here is the fix of the Order 2B to replace the 2A, which is my recommendation.
For some reasons that should be better analyzed, during the computation
of the Distributional similarity using Accelerate some parts of the
matrix are are below 1 (see the line modified by this commit). As a
consequence the log of such value below 1 gives a negative value that is
breaking the whole similarity computation: the matrix has negative and
non negative values and is becoming inconsistent.

The main fix consists in managing the values below 1 in order to get a
non negative matrix.

The Order 2A fixes the issue by adding 1 on each cell of the matrix.
The Order 2B fixes the issue by removing values below 1 only.

According to the qualitative tests I made the Graph results of order 2B
is better than the results of order 2A. Filtering seems better than
forcing using all data where some seems to be not "representative" enough.